We now recursively descend into any references clauses
and generate the UI from those tables using the tab widget.
Im surprised it worked. the current version works because i commented
out the keys which are common in both tables - but I will add a foreign
key clause to the grammar and then have the common variables factored
out automatically