Figured out how to return a table from Stored Proc
Wrote to postgresql mailing list and figured out how to return
a table. the field names in the returned table  should not be
used in the select query - otherwise they will get masked.

The stubs/postgres_sp documents my stage by stage experiments
and the mail I used to get the sp stuff sorted out.

Modified the code generator to correctly generate the sp as
required.