Basic code generation for Select done
For the simple table (no foreign key references to other tables)
I have created a Reader. This will need to be modified for
cases which have composite objects. The reader is called in a for
loop in the Get function. Not yet tested, just that the generated
code compiles. I need to delete the already created functions
and re-create them again - this is a pain as u need the function
signature in postgres to drop them - so i installed pgadmin3
but you need to have a password for the user account to connect with
that ... and i am too tired to figure that out now