Hacked till the generated code compiles
The sources need some serious cleaning up - but the generated code
compiles for now. I would not want anyone to see the source code the way
it is right now. However I have to still plough through ahead.
Next target is using the emp2.sql input file and compiling the generated
code for the case which comes up there - references MANY.

Most of the code in this checkin has to be revisited - Im ashamed of it