Fixed lot of code generation issues
At this point I need to seriously stand back and take a look at the
code - it's a serious mess - with a lot of duplication and lots
of room for lambda functions - or am i seeing room for lisp macros
here. Maybe the generator is better programmed in lisp.

however i have no respite - i have to demo the UI screens to my client
and will shift focus onto getting that demo up and ready. I have the
1st half of tomorrow to do the timesheet screen - including adding a
keyword to the grammar to make the date time selection in the timesheet a central
widget embedded in the page and not a popup display calendar