need a topological sort of tables
we need a topological sort of the tables. as the number of tables
has become high it is difficult to remember the order of creation
and dropping of tables, loading random data etc. will use this
to create scripts to create/drop tables data in the right order

this is the framework for the graph setup. i should be using boost::graph
but im in a mad rush and dont want to figure it out. im familiar with the
graph code because i did it in ncst