Fixed most code generation issues with session variables
the key point is:

	We do not want to get the entire business object for a
	session variable

All the changes in this set are code completion of the above point.
Having separate functions for _h and _cpp files has hurt me and slowed
me down again - the declarations and definitions should both be
generated from the same place, but written to different streams.

I need to fix this problem soon

Notes on this commit: The code generated still does not compile cleanly
	but I have done enough to want to have the changes stored
	securely in forge git