We are emitting a user login object on successful login
this is the key to our session management. We now need to create a
user_information object which stores to User_Login object along with
User Roles for the user. Once this is done we can start putting the
security checks in place while constructing the menu, forms and bll and
database calls.

Also note that EmployeeStatus and Exchange Details tables will not
currently work. They have to be revisited along with any tables that
have session variables as keys. We need to now start pulling session key
data from the session variable.